// src/types/ModalButtonType.ts
/**
 * Types of buttons in the modal
 * @author Gabe Abrams
 */
enum ModalButtonType {
  Okay = 'okay',
  Cancel = 'cancel',
  Yes = 'yes',
  No = 'no',
  Abandon = 'abandon',
  GoBack = 'goBack',
  Continue = 'continue',
  ImSure = 'imSure',
  Delete = 'delete',
  Confirm = 'confirm',
}

export default ModalButtonType;